% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{addCachedYAMLValue}
\alias{addCachedYAMLValue}
\title{add a stored parameter value for the named config field}
\usage{
addCachedYAMLValue(yaml = "config", field, value, subenv = NULL)
}
\arguments{
\item{yaml}{the name of the yaml container to create}

\item{field}{the parameter field to lookup}

\item{value}{to value to update}

\item{subenv}{the name of subenvironment for holding the yaml}
}
\value{
TRUE or FALSE if the change can be made
}
\description{
add a stored parameter value for the named config field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
bamFile <- system.file("extdata", "cas9_FAK76554.bam", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
addCachedYAMLValue(field="bam_file", value=bamFile)
getCachedYAMLValue(field="bam_file")

}
