% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{addCas9ParameterValue}
\alias{addCas9ParameterValue}
\title{add a stored parameter value for the named config field}
\usage{
addCas9ParameterValue(field, value)
}
\arguments{
\item{field}{the parameter field to lookup}

\item{value}{to value to update}
}
\value{
TRUE or FALSE if the change can be made
}
\description{
add a stored parameter value for the named config field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
bamFile <- system.file("extdata", "cas9_FAK76554.bam", package = "nanopoRe")
addCas9ParameterValue("bam_file", bamFile)
getCas9ParameterValue("bam_file")

}
