% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{bamSummariseByChr}
\alias{bamSummariseByChr}
\title{summarise mapping observations from a BAM file by chromosome}
\usage{
bamSummariseByChr(chrId, bamFile, force = FALSE, blockSize = 50000L)
}
\arguments{
\item{chrId}{is the chromosome to parse}

\item{bamFile}{is the location to the BAM file to parse}

\item{force}{logical value describing whether the analysis should be force recalculated}

\item{blockSize}{the number of reads to process at the time as iterating through}
}
\value{
data.frame of per read summary observations
}
\description{
This method will parse a BAM file and summarise the mapping observations in a
way that can be used for preparation of figures and confidence plots by
the nanopoRe package
}
\examples{
\dontrun{
bamSummariseByChr('MyBamFile.bam', chrId='1')
}

}
