% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{bamSummaryToCoverage}
\alias{bamSummaryToCoverage}
\title{get depth of coverage information from across the genome}
\usage{
bamSummaryToCoverage(bamFile=NULL,
    tilewidth = 1e+05,
    blocksize = 10000,
    flag = 'Primary',
    ...
)
}
\arguments{
\item{bamFile}{- path to the bamFile to use}

\item{tilewidth}{- the size of the window to use for the tiling}

\item{blocksize}{to use for parsing the BAM file}

\item{flag}{mapping type for filter (Primary/Secondary/Supplementary)}

\item{...}{such as FORCE=TRUE}
}
\value{
GRanges object with mean depth of coverage data in binned_cov field
}
\description{
This method will return a tiled Granges object containing mean depth of
coverage information
}
\examples{
demoBam <- system.file("extdata",
    "Ecoli_zymo_R10_filt_subs.bam",
    package = "nanopoRe")
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
bamSummarise(demoBam, blockSize=1000L)
bamSummaryToCoverage()

}
