% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cigar.R
\name{buildPositionMap}
\alias{buildPositionMap}
\title{prepares a matrix of operations and coordinates from a CIGAR}
\usage{
buildPositionMap(cigar)
}
\arguments{
\item{cigar}{is the CIGAR string to parse}
}
\value{
matrix of coordinates
}
\description{
This method will prepare a matrix of operations, query coordinates
and target coordinates from a Cigar string extracted from a BAM file
}
\examples{
cigarStr <-
    "46S14M5D14M2D44M3I6M2I14M3I15M1D27M1D11M1I20M1I8M2I2M3I10S"
cigarPM <- buildPositionMap(cigarStr)

}
