% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{cachedYAMLToYAML}
\alias{cachedYAMLToYAML}
\title{export the stored config parameters parsed from YAML file}
\usage{
cachedYAMLToYAML(yaml = "config", format = "YAML", subenv = NULL)
}
\arguments{
\item{yaml}{the name of the yaml container to create}

\item{format}{is the format to return results as (YAML|Kable|list)}

\item{subenv}{the name of subenvironment for holding the yaml}
}
\value{
YAML object
}
\description{
export the stored config parameters parsed from YAML file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
cachedYAMLToYAML(format="Kable")

}
