% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9CoverageTypeOverChromosomes}
\alias{cas9CoverageTypeOverChromosomes}
\title{prepare a barchart showing number of bases per chromosome coloured by mapping-type}
\usage{
cas9CoverageTypeOverChromosomes()
}
\value{
ggplot2 format barchart showing fraction of mapping types per chromosome
}
\description{
prepare a barchart showing number of bases per chromosome coloured by mapping-type;
enables looking for unexpected distributions of off-target, background etc per chromosome
}
\examples{
\dontrun{
cas9CoverageTypeOverChromosomes()
}
}
