% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9ExecutiveSummary}
\alias{cas9ExecutiveSummary}
\title{prepare executive summary infographic for the cas9 tutorial}
\usage{
cas9ExecutiveSummary()
}
\value{
path to a file that has been created
}
\description{
prepare an infographic style executive summary for key cas9 enrichment measurements
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
library(emojifont)
cas9ExecutiveSummary()

}
