% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9GetTargetList}
\alias{cas9GetTargetList}
\title{return list of defined cas9 target regions}
\usage{
cas9GetTargetList()
}
\value{
vector of target ids
}
\description{
return list of defined cas9 target regions
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9GetTargetList()

}
