% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9MappingByGenomicSegment}
\alias{cas9MappingByGenomicSegment}
\title{prepare a table of cas9 target mapping types}
\usage{
cas9MappingByGenomicSegment()
}
\value{
kable table in html format
}
\description{
the method prepares a table that defines the mapping characteristics by on-target, target-proximal, off-target and
background - parses the content produced within the broader on_tutorial_cas9 workflow
}
\examples{
\dontrun{
cas9MappingByGenomicSegment()
}
}
