% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9MultiGeneCoveragePanel}
\alias{cas9MultiGeneCoveragePanel}
\title{prepare a plot of depth-of-coverage across each of the cas9-enrichment targets included in study}
\usage{
cas9MultiGeneCoveragePanel(colMax = 4)
}
\arguments{
\item{colMax}{defines the maximum number of columns to use within the tiled matrix}
}
\value{
ggplot2 format graph
}
\description{
prepares a plot of depth-of-coverage across each of the cas9-enrichment targets included in study. This is rendered
as a multi-tile plot; simple for quick visualisation of the target regions
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9MultiGeneCoveragePanel()

}
