% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9OffTargetKaryogram}
\alias{cas9OffTargetKaryogram}
\title{prepare a karyogram shaded with regions of off-target mapping}
\usage{
cas9OffTargetKaryogram()
}
\value{
path to png format figure of plot
}
\description{
prepare a karyogram shaded with regions of off-target mapping
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9OffTargetKaryogram()

}
