% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9OffTargetTable}
\alias{cas9OffTargetTable}
\title{prepare a table of cas9 off-target mapping locations}
\usage{
cas9OffTargetTable()
}
\value{
table table
}
\description{
prepare a table of cas9 off-target mapping locations
}
\examples{
\dontrun{
cas9OffTargetTable()
}
}
