% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9ParametersToYAML}
\alias{cas9ParametersToYAML}
\title{export the stored cas9 parameters parsed from YAML file}
\usage{
cas9ParametersToYAML(format = "YAML")
}
\arguments{
\item{format}{is the format to return results as (YAML|Kable|list)}
}
\value{
YAML object
}
\description{
export the stored cas9 parameters parsed from YAML file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
cas9ParametersToYAML()
cas9ParametersToYAML("Kable")

}
