% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9SingleTargetPlot}
\alias{cas9SingleTargetPlot}
\title{prepare a plot of sequence coverage over a cas9 target region}
\usage{
cas9SingleTargetPlot(geneName, delta = 0)
}
\arguments{
\item{geneName}{is the qualuified name for the target selected}

\item{delta}{is a scaling value to zoom in beyond the target-proximal region}
}
\value{
ggplot2 figure
}
\description{
prepare a plot of depth-of-coverage over and around a specified target region
}
\examples{
\dontrun{
cas9SingleTargetPlot('HTT')
}
}
