% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9StrandedTargetPlot}
\alias{cas9StrandedTargetPlot}
\title{prepare a plot of sequence coverage shaded by strand over a cas9 target region}
\usage{
cas9StrandedTargetPlot(geneName, delta = 0)
}
\arguments{
\item{geneName}{is the qualuified name for the target selected}

\item{delta}{is a scaling value to zoom in beyond the target-proximal region}
}
\value{
ggplot2 figure
}
\description{
prepare a plot of depth-of-coverage shaded by strand over and around a specified target region
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9StrandedTargetPlot(cas9GetTargetList()[[1]])

}
