% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9TargetPerformanceTable}
\alias{cas9TargetPerformanceTable}
\title{summarise cas9 target performance}
\usage{
cas9TargetPerformanceTable()
}
\value{
kable table in html format
}
\description{
prepare tabular data showing the enrichment performance for the cas9 targets included within an enrichment study
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9TargetPerformanceTable()

}
