% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9WriteExcelOffTarget}
\alias{cas9WriteExcelOffTarget}
\title{write summary of off-target mapping results to an excel format result file}
\usage{
cas9WriteExcelOffTarget()
}
\value{
None
}
\description{
write summary of off-target mapping results to an excel format result file
}
\examples{
\dontrun{
cas9WriteExcelOffTarget()
}
}
