% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9WriteExcelOnTarget}
\alias{cas9WriteExcelOnTarget}
\title{write summary of on-target mapping results to an excel format result file}
\usage{
cas9WriteExcelOnTarget()
}
\value{
path to file created
}
\description{
write summary of on-target mapping results to an excel format result file
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
importCas9TutorialData(system.file("extdata", package="nanopoRe"))
cas9WriteExcelOnTarget()

}
