% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentExecutiveSummary}
\alias{enrichmentExecutiveSummary}
\title{prepare executive summary infographic for the cas9 tutorial}
\usage{
enrichmentExecutiveSummary()
}
\value{
path to a file that has been created
}
\description{
prepare an infographic style executive summary for key cas9 enrichment
measurements
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
library(emojifont)
enrichmentExecutiveSummary()

}
