% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentGetTargetList}
\alias{enrichmentGetTargetList}
\title{return list of defined cas9 target regions}
\usage{
enrichmentGetTargetList()
}
\value{
vector of target ids
}
\description{
return list of defined cas9 target regions
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentGetTargetList()

}
