% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentMappingByGenomicSegment}
\alias{enrichmentMappingByGenomicSegment}
\title{prepare a table of cas9 target mapping types}
\usage{
enrichmentMappingByGenomicSegment()
}
\value{
kable table in html format
}
\description{
the method prepares a table that defines the mapping characteristics by
on-target, target-proximal, off-target and background - parses the content
produced within the broader on_tutorial_cas9 workflow
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentMappingByGenomicSegment()

}
