% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentOffTargetKaryogram}
\alias{enrichmentOffTargetKaryogram}
\title{prepare a karyogram shaded with regions of off-target mapping}
\usage{
enrichmentOffTargetKaryogram(fill = FALSE)
}
\arguments{
\item{fill}{- if TRUE will populate chromosome lengths and ids for human}
}
\value{
path to png format figure of plot
}
\description{
prepare a karyogram shaded with regions of off-target mapping
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentOffTargetKaryogram()

}
