% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentOffTargetTable}
\alias{enrichmentOffTargetTable}
\title{prepare a table of cas9 off-target mapping locations}
\usage{
enrichmentOffTargetTable(maxRows = 10)
}
\arguments{
\item{maxRows}{the maximum number of rows to display in the table}
}
\value{
table table
}
\description{
prepare a table of cas9 off-target mapping locations
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentOffTargetTable()

}
