% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentStrandedTargetPlot}
\alias{enrichmentStrandedTargetPlot}
\title{prepare a plot of sequence coverage shaded by strand over a cas9 target region}
\usage{
enrichmentStrandedTargetPlot(geneName, delta = 0)
}
\arguments{
\item{geneName}{is the qualuified name for the target selected}

\item{delta}{is a scaling value to zoom in beyond the target-proximal region}
}
\value{
ggplot2 figure
}
\description{
prepare a plot of depth-of-coverage shaded by strand over and around a specified target region
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentStrandedTargetPlot(enrichmentGetTargetList()[[1]])

}
