% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{enrichmentWriteExcelOnTarget}
\alias{enrichmentWriteExcelOnTarget}
\title{write summary of on-target mapping results to an excel format result file}
\usage{
enrichmentWriteExcelOnTarget()
}
\value{
path to file created
}
\description{
write summary of on-target mapping results to an excel format result file
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
importEnrichmentResults(system.file("extdata", package="nanopoRe"))
enrichmentWriteExcelOnTarget()

}
