% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{extract5mC}
\alias{extract5mC}
\title{prepares a data.frame of per-read methylC bases from provided fast5 file}
\usage{
extract5mC(fast5file, threshold_5mc = 0.85, fast5files = NULL, ...)
}
\arguments{
\item{fast5file}{is the fast5 file to parse}

\item{threshold_5mc}{is the cutoff to apply to values returned}

\item{fast5files}{is a vector of files being processed (for pretty logging)}

\item{...}{beyond}
}
\value{
data.frame of coordinates and sequence names from fast5
}
\description{
This method will parse a fast5 file to extract the 5mC bases classified as 5mC with a
probability >= the provided threshold_5mc variable
}
