% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{extract5mCProbabilities}
\alias{extract5mCProbabilities}
\title{prepares a data.frame of per-read methylC probability counts}
\usage{
extract5mCProbabilities(fast5file, fast5files = NULL, force = FALSE,
  b = 33, ...)
}
\arguments{
\item{fast5file}{is the fast5 file to parse}

\item{fast5files}{is a vector of files being processed (for pretty logging)}

\item{force}{to force recalculation}

\item{b}{the number of sample breaks to make}

\item{...}{beyond}
}
\value{
data.frame of coordinates and sequence names from fast5
}
\description{
This method will parse a fast5 file to the distribution of probabilities for C residues
actually being 5mC residues
}
