% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{extractMethylatedBases}
\alias{extractMethylatedBases}
\title{prepares a data.frame of read mapping context for methylation parsing}
\usage{
extractMethylatedBases(x, filteredChunk, modifications_df)
}
\arguments{
\item{x}{is a numeric pointer}

\item{filteredChunk}{corresponds to parsed BAM file data.frame}

\item{modifications_df}{corresponds to parsed methylation data.frame}
}
\value{
data.frame
}
\description{
This method will parses the specified BAM file to extract CIGAR and other coordinates for
the purpose of selecting sequence bases that are base modified
}
