% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastqValidator}
\alias{fastqValidator}
\title{parse a fastq file aiming to validate sequences}
\usage{
fastqValidator(fastq)
}
\arguments{
\item{fastq}{A fastq format DNA/RNA sequence file}
}
\value{
logical defining if fastq provided is indeed valid fastq
}
\description{
parse a fastq file aiming to validate sequences
}
