% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixFastq}
\alias{fixFastq}
\title{fix a corrupted fastq file (if fastq-like)}
\usage{
fixFastq(fastq, newfastq)
}
\arguments{
\item{fastq}{file location of fastq source}

\item{newfastq}{location of file to write content to}
}
\value{
path to new fastq file (same as newfastq provided)
}
\description{
fixFastq parses a fastq (or fastq.gz) file for compliant fastq entries and writes these to
the file specified in newfastq parameter. Any non-compliant reads are dropped
}
\examples{
badFastq <- system.file("extdata", "frankenFastq.fastq.gz", package = "nanopoRe")
fastqValidator(badFastq)
tempFile <- tempfile(pattern="fastq", fileext=".fq")
fixFastq(badFastq, tempFile)
fastqValidator(tempFile)

}
