% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{getCachedYAMLFields}
\alias{getCachedYAMLFields}
\title{list the parameter fields stored in a cached YAML config file}
\usage{
getCachedYAMLFields(yaml = "config", subenv = NULL)
}
\arguments{
\item{yaml}{the name of the YAML object}

\item{subenv}{the sub-environment to process}
}
\value{
vector of field names
}
\description{
list the parameter fields stored in a cached YAML config file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
getCachedYAMLFields()

}
