% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{getCachedYAMLValue}
\alias{getCachedYAMLValue}
\title{get the stored parameter value from named field}
\usage{
getCachedYAMLValue(yaml = "config", field, subenv = NULL)
}
\arguments{
\item{yaml}{the name of the YAML object}

\item{field}{the parameter field to lookup}

\item{subenv}{the sub-environment to process}
}
\value{
value from config
}
\description{
get the stored parameter value from named field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
getCachedYAMLValue(field="fastq")

}
