% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{getCas9ParameterFields}
\alias{getCas9ParameterFields}
\title{list the parameter fields stored with the YAML config file}
\usage{
getCas9ParameterFields()
}
\value{
vector of field names
}
\description{
list the parameter fields stored with the YAML config file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
getCas9ParameterFields()

}
