% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{getCas9ParameterValue}
\alias{getCas9ParameterValue}
\title{get the stored parameter value from named field}
\usage{
getCas9ParameterValue(field)
}
\arguments{
\item{field}{the parameter field to lookup}
}
\value{
value from config
}
\description{
get the stored parameter value from named field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
getCas9ParameterValue("fastq")

}
