% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getChromosomeIds}
\alias{getChromosomeIds}
\title{get chromosome identifiers from reference genome}
\usage{
getChromosomeIds()
}
\value{
vector of names
}
\description{
This method will return a vector of chromsome identifiers
}
\examples{
init()
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
getChromosomeIds()

}
