% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getChromosomeSequence}
\alias{getChromosomeSequence}
\title{returns a DNAStringSet object corresponding to specified chromosome from reference genome}
\usage{
getChromosomeSequence(dnaStringSetId)
}
\arguments{
\item{dnaStringSetId}{is the pointer to use}
}
\value{
DNAStringSet corresponding to pointer provided
}
\description{
This method will return a DNAStringSet object from the reference genome; requires a numeric pointer to the object
}
\examples{
\dontrun{
getChromosomeSequence(1)
}

}
\seealso{
[getStringSetId()] for method to prepare numeric pointer
}
