% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFastqBases}
\alias{getFastqBases}
\title{return the number of fastq bases previously parsed from
provided Fastq file}
\usage{
getFastqBases()
}
\value{
long long integer of read fastq bases
}
\description{
return the number of fastq bases previously parsed from
provided Fastq file
}
\examples{
fastq <- system.file("extdata", "example.fastq.gz", package = "nanopoRe")
fastqValidator(fastq)
getFastqBases()

}
