% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFastqCount}
\alias{getFastqCount}
\title{return the number of fastq entries previously parsed from provided Fastq file}
\usage{
getFastqCount()
}
\value{
long integer of read fastq elements
}
\description{
return the number of fastq entries previously parsed from provided Fastq file
}
\examples{
fastq <- system.file("extdata", "example.fastq.gz", package = "nanopoRe")
fastqValidator(fastq)
getFastqCount()

}
