% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFastqPlusErrorCount}
\alias{getFastqPlusErrorCount}
\title{count of fastq elements rejected due to line 3 plus separator}
\usage{
getFastqPlusErrorCount()
}
\value{
integer count of malformed fastq '+' separator entries
}
\description{
count of fastq elements rejected due to line 3 plus separator
}
\examples{
fastq <- system.file("extdata", "frankenFastq.fastq.gz", package = "nanopoRe")
fastqValidator(fastq)
getFastqPlusErrorCount()

}
