% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowcellLayout.R
\name{getFlongleChannelMap}
\alias{getFlongleChannelMap}
\title{produce the channelMap for a flongle flowcell for spatial plots}
\usage{
getFlongleChannelMap()
}
\value{
data.frame with channel, row and col columns
}
\description{
prepares a data.frame of channelIds and their X, Y coordinates for a flongle flowcell
}
\examples{
head(getFlongleChannelMap())

}
