% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{getGitDescription}
\alias{getGitDescription}
\title{get git metadata for current project}
\usage{
getGitDescription()
}
\value{
kable
}
\description{
The ONT tutorials are typically checked out of github - there is a potential
for misalignment of a user's working branch and current - would be sensible
to include a brief git comment so that Techical Support can follow-up with
assessment of versioning. This code will not work with code that has been
extracted from either CONDA or Bioconductor ...
}
\examples{
init()
getGitDescription()

}
