% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMalformedFastqHeaderCount}
\alias{getMalformedFastqHeaderCount}
\title{count of fastq elements rejected due to malformed fastq header}
\usage{
getMalformedFastqHeaderCount()
}
\value{
integer count of malformed fastq header entries
}
\description{
count of fastq elements rejected due to malformed fastq header
}
\examples{
fastq <- system.file("extdata", "frankenFastq.fastq.gz", package = "nanopoRe")
fastqValidator(fastq)
getMalformedFastqHeaderCount()

}
