% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowcellLayout.R
\name{getMinIONChannelMap}
\alias{getMinIONChannelMap}
\title{produce the channelMap for a MinION flowcell for spatial plots}
\usage{
getMinIONChannelMap()
}
\value{
data.frame with channel, row and col columns
}
\description{
prepares a data.frame of channelIds and their X, Y coordinates for a MinION flowcell
}
\examples{
head(getMinIONChannelMap())

}
