% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{getOutputFormat}
\alias{getOutputFormat}
\title{get the defined output format for ggplot2 figures}
\usage{
getOutputFormat()
}
\value{
character of format
}
\description{
This method defines the expected output format for figure management
}
\examples{
getOutputFormat()
setOutputFormat(listOutputFormat()[2])
getOutputFormat()

}
