% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{getPlotDimensions}
\alias{getPlotDimensions}
\title{return defined dimensions for an allowed ggplot2 object}
\usage{
getPlotDimensions()
}
\value{
list defining dpi, width, height and units
}
\description{
This method presents the list with the defined metrics for a ggplot2 presentation
}
\examples{
getPlotDimensions()

}
