% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowcellLayout.R
\name{getPromethIONChannelMap}
\alias{getPromethIONChannelMap}
\title{produce the channelMap for a PromethION flowcell for spatial plots}
\usage{
getPromethIONChannelMap()
}
\value{
data.frame with channel, row and col columns
}
\description{
prepares a data.frame of channelIds and their X, Y coordinates for a
PromethION flowcell
}
\examples{
head(getPromethIONChannelMap())

}
