% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getReferenceGenome}
\alias{getReferenceGenome}
\title{get path reference genome object}
\usage{
getReferenceGenome()
}
\value{
character representation of path to stored fasta format file object
}
\description{
This method returns the stored path for the reference genome object
}
\examples{
\dontrun{
getReferenceGenome()
}

}
