% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getReferenceGenomeGR}
\alias{getReferenceGenomeGR}
\title{returns a GRanges object corresponding to the currently loaded refgenome}
\usage{
getReferenceGenomeGR()
}
\value{
DNAStringSet corresponding to pointer provided
}
\description{
This method will return a DNAStringSet object from the reference genome;
requires a numeric pointer to the object
}
\examples{
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
getReferenceGenomeGR()

}
