% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{getRpath}
\alias{getRpath}
\title{get path for R results}
\usage{
getRpath()
}
\value{
character representation of path
}
\description{
This method returns the path where default R results will be stored during an analysis
}
\examples{
\dontrun{
getRpath()
}

}
