% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{get chromosome lengths for a named set of chromosomes}
\usage{
getSeqLengths(x)
}
\arguments{
\item{x}{vector of chromosome ids}
}
\value{
vector of names
}
\description{
This method will return a vector of named chromosome lengths
}
\examples{
init()
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
getSeqLengths(getChromosomeIds())

}
