% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSequenceQualityMismatchCount}
\alias{getSequenceQualityMismatchCount}
\title{count of fastq elements rejected due to mismatch between sequence and quality field lengths}
\usage{
getSequenceQualityMismatchCount()
}
\value{
integer count of fastq entries with seq/qual length challenges
}
\description{
count of fastq elements rejected due to mismatch between sequence and quality field lengths
}
