% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSkippedLineCount}
\alias{getSkippedLineCount}
\title{count of lines of fastq file skipped to enable fastq entry parsing}
\usage{
getSkippedLineCount()
}
\value{
integer count of lines skipped
}
\description{
count of lines of fastq file skipped to enable fastq entry parsing
}
\examples{
fastq <- system.file("extdata", "frankenFastq.fastq.gz",
    package = "nanopoRe")
fastqValidator(fastq)
getSkippedLineCount()

}
