% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getStringSetId}
\alias{getStringSetId}
\title{accessory method for mapping named chromosomes to their pointers in the reference fasta}
\usage{
getStringSetId(chrId)
}
\arguments{
\item{chrId}{is the identifier for the chromosomes to be dropped}
}
\value{
None
}
\description{
This method will return the numerical index for the named chromosome within the referenceGenomeSequence
}
\examples{
\dontrun{
getStringSetId("1")
}

}
