% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getZeroLengthSequenceCount}
\alias{getZeroLengthSequenceCount}
\title{count of fastq elements rejected due to zero length sequence}
\usage{
getZeroLengthSequenceCount()
}
\value{
integer count of fastq entries with zero sequence length
}
\description{
count of fastq elements rejected due to zero length sequence
}
