% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPresentation.R
\name{ggbiosave}
\alias{ggbiosave}
\title{save a ggbio plot to file}
\usage{
ggbiosave(filename,
    plot = last_plot(), device = default_device(filename), path = NULL,
    scale = 1, width = par('din')[1], height = par('din')[2],
    units = c('in', 'cm', 'mm'), dpi = 300, limitsize = TRUE, ...)
}
\arguments{
\item{filename}{filename}

\item{plot}{plot}

\item{device}{device}

\item{path}{path}

\item{scale}{scale}

\item{width}{width}

\item{height}{height}

\item{units}{units}

\item{dpi}{dpi}

\item{limitsize}{limitsize}

\item{...}{to downstream methods}
}
\value{
path to image
}
\description{
save a ggbio plot to file
}
