% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedMethods.R
\name{harvestBam}
\alias{harvestBam}
\title{parse (indexed) BAM file for summary data for a given windows on a specific chromosome}
\usage{
harvestBam(x, dnaStringSetId, chrId, window.size, bamfilelocation,
  level = "PRIMARY")
}
\arguments{
\item{x}{is a numeric pointer for the interval start position}

\item{dnaStringSetId}{is the chromosome pointer within the reference genome}

\item{chrId}{is the name of the chromosome of interest}

\item{window.size}{is the width of the window in nucleotides}

\item{bamfilelocation}{is the location to the BAM file to parse}

\item{level}{is one of (c("PRIMARY", "SECONDARY", "SUPPLEMENTARY"))}
}
\value{
vector of summary observations
}
\description{
This method will direct the parsing of a specified BAM file for a named chromsome and qualified interval
}
\examples{
\dontrun{
harvestBam(100000, 1, "1", 100000, file.path("Analysis", "Minimap2", "MyBamFile.bam"))
}

}
