% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{harvestChromosome}
\alias{harvestChromosome}
\title{parse (indexed) BAM file for summary data for specific chromosome}
\usage{
harvestChromosome(chrId, bamfilelocation, window.size = 1e+05,
  mc.cores = min(detectCores() - 1, 24), force = FALSE)
}
\arguments{
\item{chrId}{is the name of the chromosome of interest}

\item{bamfilelocation}{is the location to the BAM file to parse}

\item{window.size}{is the size of the stepping window to apply (default 100000)}

\item{mc.cores}{defines the number of cores to use (default = n-1 of available cores)}

\item{force}{is a logical defining whether the analysis should be recalculated (cached version will be reused if possible)}
}
\value{
data.frame of summary observations
}
\description{
This method will direct the parsing of a specified BAM file for a named chromsome
}
\examples{
\dontrun{
harvestChromosome("1", file.path("Analysis", "Minimap2", "MyBamFile.bam"))
}

}
