% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{harvestUnmappedReads}
\alias{harvestUnmappedReads}
\title{extract unmapped read quality information from provided BAM file}
\usage{
harvestUnmappedReads(bamfilelocation, force = FALSE)
}
\arguments{
\item{bamfilelocation}{is the location to the BAM file to parse}

\item{force}{is a boolean defining whether an existing result file should be overwritten}
}
\value{
vector of summary observations
}
\description{
This method will parse the unmapped reads from a BAM file and return a data frame containing
readID and quality information
}
\examples{
\dontrun{
harvestUnmappedReads(file.path("Analysis", "Minimap2", "MyBamFile.bam"))
}

}
