% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{hasCachedYAMLField}
\alias{hasCachedYAMLField}
\title{check for available named cas9 parameter field}
\usage{
hasCachedYAMLField(yaml = "config", field, subenv = NULL)
}
\arguments{
\item{yaml}{the name of the YAML object}

\item{field}{is the name of a parameter field}

\item{subenv}{the sub-environment to process}
}
\value{
TRUE or FALSE
}
\description{
check for available named cas9 parameter field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
hasCachedYAMLField(field="bam_file")
hasCachedYAMLField(field="study_name")

}
