% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{hasCas9ParameterField}
\alias{hasCas9ParameterField}
\title{check for available named cas9 parameter field}
\usage{
hasCas9ParameterField(field)
}
\arguments{
\item{field}{is the name of a parameter field}
}
\value{
TRUE or FALSE
}
\description{
check for available named cas9 parameter field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
hasCas9ParameterField("fastq")
hasCas9ParameterField("bam_file")

}
