% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{importConfigYAML}
\alias{importConfigYAML}
\title{import the cas9 parameters from YAML file}
\usage{
importConfigYAML(yaml = "config", yamlFile, subenv = NULL)
}
\arguments{
\item{yaml}{the name of the yaml container to create}

\item{yamlFile}{is a path to a stored YAML config file}

\item{subenv}{the name of subenvironment for holding the yaml}
}
\value{
None
}
\description{
import the cas9 parameters from YAML file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)

}
