% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{importSequencingSummary}
\alias{importSequencingSummary}
\title{load a sequencing_summary.txt file into memory}
\usage{
importSequencingSummary(seqsum, cache = TRUE, chunksize = 1e+06)
}
\arguments{
\item{seqsum}{is a path to a file}

\item{cache}{whether the object created should be cached in env}

\item{chunksize}{the number of reads to parse per iteration block}
}
\value{
data.frame of observations from the sequencing_summary.txt file
provided
}
\description{
importSequencingSummary loads a sequencing_summary.txt file into
memory and performs basic sanity checking to ensure that the file
is cleaned of potential duplicate headers from aggregation
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)

}
