% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfographicPlot.R
\name{infoGraphicPlot3}
\alias{infoGraphicPlot3}
\title{prepare an 3x infographic-styled summary plot}
\usage{
infoGraphicPlot3(identifier, panelA, panelB, panelC, reportDPI = 90)
}
\arguments{
\item{identifier}{a key for the plot}

\item{panelA}{information for panel1 in format of panelA=c(key="", value="", icon="")}

\item{panelB}{see panelA}

\item{panelC}{see panelA}

\item{reportDPI}{resolution for plot}
}
\value{
link to a figure for display
}
\description{
This method will prepare a 3x graphical plot summarising user data with an emphasis on visualisation
}
\examples{
\dontrun{
infoGraphicPlot3(identifier="mappingSummary",
    panelA=c(key="Alignments", value="123456", icon="fa-sliders"),
    panelB=c(key="Mapping Yield", value="680M", icon="fa-calculator"),
    panelC=c(key="Average Accuracy", value="89.0 \%", icon="fa-area-chart"))
}
}
