% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{init}
\alias{init}
\title{Initialise the nanopoRe environment}
\usage{
init()
}
\value{
None
}
\description{
Creates a nanopoRe environment; package specific parameters and values will be stored within this
environment; the name of the environment is defined internally
}
\examples{
\dontrun{
init()
}
}
