% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{is_casDataRun}
\alias{is_casDataRun}
\title{accessory method to test whether the cas9 data indexing has been performed}
\usage{
is_casDataRun(resultsDir = NULL)
}
\arguments{
\item{resultsDir}{the location of the folder where files should be pulled from - this should
be determined automatically from config.yaml in most cases}
}
\value{
TRUE or FALSE depending
}
\description{
The cas9 workflow involves a deep dive into the experimental data and produces
a couple of Rdata files that contain the experimental context and information
relating to on-target, off-target and other effects. This method tests whether
the analysis has been performed
}
\examples{
sourceCas9Parameters(system.file("extdata", "cas9_demo.yaml", package = "nanopoRe"))
# some demo data is stored with the package - import the data by supply the non-standard
# path
is_casDataRun(system.file("extdata", package="nanopoRe"))

}
