% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{isInitialised}
\alias{isInitialised}
\title{check NanopoRe environment}
\usage{
isInitialised()
}
\value{
a logical defining whether the environment is initialised
}
\description{
performs a sanity check to ensure that NanopoRe environment is initialised
}
