% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceCharacteristics.R
\name{lcalc}
\alias{lcalc}
\title{calculate Lstatistics (e.g. L50) for given sequence collection}
\usage{
lcalc(len.vector, n)
}
\arguments{
\item{len.vector}{is a vector of sequence lengths}

\item{n}{is a numeric of the value to use}
}
\value{
character representation of md5sum
}
\description{
Given a set of contigs, each with its own length, the L50 count is defined
as the smallest number of contigs whose length sum makes up half of genome
size (https://en.wikipedia.org/wiki/N50,_L50,_and_related_statistics)
}
\examples{
lcalc(seq(1, 10), 0.5)

}
