% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{listCachedObjects}
\alias{listCachedObjects}
\title{list objects that are stored within the nanopoRe working space}
\usage{
listCachedObjects(subenv = NULL)
}
\arguments{
\item{subenv}{the nested environment to consider}
}
\value{
character vector of objects stored in the nanopoRe environment
}
\description{
The nanopoRe package makes use of R environments to store data in memory,
but within an internally-named environment; this method dumps out (for
reasons of QC and transparency) a complete listing of these stored objects
}
\examples{
cachedObjects <- listCachedObjects()

}
