% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{listOutputFormat}
\alias{listOutputFormat}
\title{list available output frameworks for ggplot2 figures}
\usage{
listOutputFormat()
}
\value{
vector of allowed types
}
\description{
This method presents the list of allowed data types for managing the
simplified presentation of ggplot2 figures
}
\examples{
listOutputFormat()

}
