% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast5.R
\name{loadMethylationBamFile}
\alias{loadMethylationBamFile}
\title{prepares a data.frame of read mapping context for methylation parsing}
\usage{
loadMethylationBamFile(bamFile, chrId, force = FALSE)
}
\arguments{
\item{bamFile}{is the bam file to parse}

\item{chrId}{is the identifier for the chromosome to parse}

\item{force}{is logical specifying whether the load should be forced (recalculated)}
}
\value{
data.frame
}
\description{
This method will parses the specified BAM file to extract CIGAR and other coordinates for
the purpose of selecting sequence bases that are base modified
}
