% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{loadReferenceGenome}
\alias{loadReferenceGenome}
\title{load reference genome into memory}
\usage{
loadReferenceGenome()
}
\value{
None
}
\description{
This method will parse the defined reference.file into a DNAStringSet object for handling in memory
}
\examples{
\dontrun{
getReferenceGenome()
}

}
