% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{parallelBamSummarise}
\alias{parallelBamSummarise}
\title{summarise mapping observations from a BAM file using parallel chr-by-chr}
\usage{
parallelBamSummarise(bamFile, force = FALSE,
  mc.cores = min(parallel::detectCores() - 1, 24))
}
\arguments{
\item{bamFile}{is the location to the BAM file to parse}

\item{force}{logical value describing whether the analysis should be force recalculated}

\item{mc.cores}{number of threads to use for the process}
}
\value{
data.frame of per read summary observations
}
\description{
This method will parse a BAM file and summarise the mapping observations in a
way that can be used for preparation of figures and confidence plots by
the nanopoRe package
}
\examples{
\dontrun{
parallelBamSummarise(file.path('Analysis', 'Minimap2', 'MyBamFile.bam'))
}

}
