% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecatedMethods.R
\name{parseBamFile}
\alias{parseBamFile}
\title{parse BAM file to calculate summary statistics}
\usage{
parseBamFile(bamfilelocation, window.size = 1e+05,
  mc.cores = min(detectCores() - 1, 24), force = FALSE,
  level = "PRIMARY")
}
\arguments{
\item{bamfilelocation}{is the location to the BAM file to parse}

\item{window.size}{is the size of the stepping window to apply (default 100000)}

\item{mc.cores}{defines the number of cores to use (default = n-1 of available cores)}

\item{force}{is a logical defining whether the analysis should be recalculated (cached version will be reused if possible)}

\item{level}{is one of (c("PRIMARY", "SECONDARY", "SUPPLEMENTARY"))}
}
\value{
data.frame of summary observations
}
\description{
This method will direct the parsing of a BAM file to extract summary mapping statistics
}
\examples{
\dontrun{
parseBamFile(file.path("Analysis", "Minimap2", "MyBamFile.bam"))
}

}
