% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phred.R
\name{phredmean}
\alias{phredmean}
\title{calculate mean Phred scores from list of Q values}
\usage{
phredmean(q)
}
\arguments{
\item{q}{is a vector of q values}
}
\value{
mean phred scaled q-value
}
\description{
This method calculates the mean quality value from a list of q-values; handles the
challenge of relative log-scaling of the Q and challenges of a more linear mean; this
avoids an artificial inflation of phred-scores
}
\examples{
mean(c(20,30,40))
phredmean(c(20,30,40))

}
