% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentPlots.R
\name{plotAlignmentAccuracy}
\alias{plotAlignmentAccuracy}
\title{plot a histogram of mapping accuracies}
\usage{
plotAlignmentAccuracy(bamFile, flag = "Primary", lower = 0.7,
    upper = 1, segments = 50)
}
\arguments{
\item{bamFile}{is the path to the bamFile to plot}

\item{flag}{is the flagged sequence type to plot [Primary]}

\item{lower}{lower limit for plotting values}

\item{upper}{upper limit for plotting values}

\item{segments}{number of segments to split data into}
}
\value{
a ggplot2 object
}
\description{
This method will plot a histogram of mapping accuracies
}
\examples{
\dontrun{
plotAlignmentAccuracy(bamFile)
}

}
