% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsDepthOfCoverage.R
\name{plotDepthOfCoverageMegablock}
\alias{plotDepthOfCoverageMegablock}
\title{plot a tiled panel of chromosomal depths-of-coverage}
\usage{
plotDepthOfCoverageMegablock(parsedBamFile, colMax = 4)
}
\arguments{
\item{parsedBamFile}{is the location to the BAM file to parse}

\item{colMax}{defines the number of columns}
}
\description{
This method will display depths-of-coverage for the chromosomes contained in the provided parsedBamFile
container
}
\examples{
\dontrun{
plotDepthOfCoverageMegablock(parsedBamFile)
}

}
