% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsDepthOfCoverage.R
\name{plotOverallCovHistogram}
\alias{plotOverallCovHistogram}
\title{plot a histogram of whole genome depth-of-coverage}
\usage{
plotOverallCovHistogram(bamFile)
}
\arguments{
\item{bamFile}{is the path to the bamFile to plot}
}
\description{
This method will plot a histogram of whole genome depth-of-coverage
}
\examples{
\dontrun{
plotDepthOfCoverageMegablock(bamFile)
}

}
