% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phred.R
\name{qualToMeanQ}
\alias{qualToMeanQ}
\title{calculate mean Phred score from an ASCII encoded phred string}
\usage{
qualToMeanQ(qstr)
}
\arguments{
\item{qstr}{is an ASCII encoded Phred quality score}
}
\value{
mean phred scaled q-value
}
\description{
FASTQ and BAM store per base qualities as an ASCII string. Accessory methods in e.g. ShortRead allow for a
sum of the numeric encoded scores; this is not corrected for the log/linear so scores are synthetically boosted
- this simple method performs mean on the character level data ...
}
\examples{
 \dontrun{
qualToMeanQ('ABCDEF')
}

}
