% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetEnrichment.R
\name{RunEnrichmentAnalysis}
\alias{RunEnrichmentAnalysis}
\title{perform an R-based enrichment analysis for a target region of the genome}
\usage{
RunEnrichmentAnalysis(force = FALSE, ...)
}
\arguments{
\item{force}{defines whether a reanalysis of the data should be forced
if results already existing}

\item{...}{- who knows - I'm using this for mc.cores}
}
\value{
TRUE or FALSE depending
}
\description{
This method is associated with the cas9-mediated enrichment of genomic
target regions ...
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
is_casDataRun()
# this is a header to a document - not a live Snakemake workspace - demo
# how to actually run ad-hoc the code
setCas9ParameterValue(
    "reference_genome", system.file("extdata", "cas9_demo_ref.fasta", package = "nanopoRe"))
setCas9ParameterValue(
    "target_regions", system.file("extdata", "cas9_demo_target.bed", package = "nanopoRe"))
addCas9ParameterValue(
    "bam_file", system.file("extdata", "cas9_FAK76554.bam", package = "nanopoRe"))
addCas9ParameterValue(
    "unmapped_quals", system.file("extdata", "cas9_FAK76554.unmapped.quals", package = "nanopoRe"))
# lines between the previous comment and here are typically not required
RunEnrichmentAnalysis(mc.cores=1)

}
