% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequenceSummaryBarcodeHistogram}
\alias{SequenceSummaryBarcodeHistogram}
\title{present a histogram of read count by sorted barcode Id}
\usage{
SequenceSummaryBarcodeHistogram(seqsum, bctheshold = 150)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{bctheshold}{the threshold number of reads for a barcode to be considered (150)}
}
\value{
a ggplot2 histogram or NULL
}
\description{
present a ggplot2 histogram of read counts by sorted barcode Id
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
SequenceSummaryBarcodeHistogram(seqsum)
}

}
