% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequenceSummaryBarcodeTable}
\alias{SequenceSummaryBarcodeTable}
\title{tabulate information on the SequencingSummary barcode fields and status}
\usage{
SequenceSummaryBarcodeTable(seqsum = NA, bcthreshold = 150)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{bcthreshold}{the threshold number of reads for a barcode to be considered (150)}
}
\value{
a kable table or NULL
}
\description{
tabulate information on the SequencingSummary barcode fields and status
}
\examples{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
importSequencingSummary(seqsumFile)
barcodeFile <- system.file('extdata', 'barcoding_summary.txt.bz2', package = 'nanopoRe')
SequencingSummaryBarcodeMerge(barcodeFile=barcodeFile)
SequenceSummaryBarcodeTable()

}
