% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequenceSummaryBasicInfoPlot}
\alias{SequenceSummaryBasicInfoPlot}
\title{present an infographic styled basic characteristics plot of sequence_summary.txt content}
\usage{
SequenceSummaryBasicInfoPlot(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
file path to ggplot2 format file
}
\description{
present an infographic styled basic characteristics plot of sequence_summary.txt content
}
\examples{
\dontrun{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequenceSummaryBasicInfoPlot(seqsum)
}

}
