% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequenceSummaryExecutiveSummary}
\alias{SequenceSummaryExecutiveSummary}
\title{present an infographic styled executive summary of sequence_summary.txt content}
\usage{
SequenceSummaryExecutiveSummary(seqsum, flowcellId = "undefined")
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{flowcellId}{is a label for the plot}
}
\value{
file path to ggplot2 format file
}
\description{
present an infographic styled executive summary of sequence_summary.txt content
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequenceSummaryExecutiveSummary(seqsum)
}

}
