% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryActiveChannelPlot}
\alias{SequencingSummaryActiveChannelPlot}
\title{plot number of observed channels actively producing data against time}
\usage{
SequencingSummaryActiveChannelPlot(seqsum,
    scaling = 1,
    sampleHours = 48,
    sampleIntervalMinutes = 60
)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{scaling}{scale factor for the data}

\item{sampleHours}{is the number of hours to plot data for (default is 48)}

\item{sampleIntervalMinutes}{is the resolution to plot data at}
}
\value{
ggplot2 showing temporal performance
}
\description{
plots a ggplot2 plot of active channels against time
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryActiveChannelPlot(seqsum)

}
