% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequencingSummaryBarcodeMerge}
\alias{SequencingSummaryBarcodeMerge}
\title{merge sequencing summary and barcode summary files}
\usage{
SequencingSummaryBarcodeMerge(seqsum = NA, barcodeFile = NULL)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{barcodeFile}{pointer to a barcode file as produced by Guppy}
}
\value{
a data.frame with barcode information
}
\description{
This method is used by the BasicQC tutorial to merge the common sequence
identifiers shared between sequencing_summary and barcode_summary
}
\examples{
seqsumFile <- system.file(
    'extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
importSequencingSummary(seqsumFile)
barcodeFile <- system.file(
    'extdata', 'barcoding_summary.txt.bz2', package = 'nanopoRe')
SequencingSummaryBarcodeMerge(barcodeFile=barcodeFile)

}
