% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryBase50}
\alias{SequencingSummaryBase50}
\title{calculates the fractional number of bases according to supplied b parameter}
\usage{
SequencingSummaryBase50(seqsum, b = 0.5)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{b}{is a fractional point through run against which time will be calculated}
}
\value{
a numeric value expressed in gigabases
}
\description{
an accessory method for various logicals; simple fractional base calculator
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
Base50 <- SequencingSummaryBase50(seqsum)
}

}
