% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryChannelActivity}
\alias{SequencingSummaryChannelActivity}
\title{prepare a channel activity plot from sequencing_summary reads file}
\usage{
SequencingSummaryChannelActivity(seqsum = NA, platform = NA)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{platform}{is the nanopore platform [MinION/Flongle/PromethION]}
}
\value{
ggplot2 channel activity plot
}
\description{
plots the basic eye-candy gauge channel activity plot of reads against channel of origin
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryChannelActivity(seqsum)

}
