% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryCumulativeBases}
\alias{SequencingSummaryCumulativeBases}
\title{plot cumulative volumes of sequence bases}
\usage{
SequencingSummaryCumulativeBases(seqsum,
    sampleHours = NA,
    sampleIntervalMinutes = 60
)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{sampleHours}{is the number of hours to plot data for}

\item{sampleIntervalMinutes}{is the resolution to plot data at}
}
\value{
ggplot2 showing temporal performance
}
\description{
plots a ggplot2 graph of accumulated sequenced bases against time for run
and separates bases called from passed and failed sequence reads
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryCumulativeBases(seqsum)

}
