% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryExtractRuntime}
\alias{SequencingSummaryExtractRuntime}
\title{from sequencing_summary.txt return integer describing runtime in hours}
\usage{
SequencingSummaryExtractRuntime(seqsum = NA)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
integer of runtime in hours
}
\description{
from provided sequencing_summary.txt file return an integer best describing
the runtime in hours
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
importSequencingSummary(seqsumFile)
SequencingSummaryExtractRuntime()

}
