% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryFlowCellID}
\alias{SequencingSummaryFlowCellID}
\title{get flowcell id from sequencing_summary file}
\usage{
SequencingSummaryFlowCellID(seqsumFile)
}
\value{
character vector of flowcell id
}
\description{
The sequencing summary file contains a load of additional information;
this method will quickly extract and return the flowcell id
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
FCid <- SequencingSummaryFlowCellID(seqsumFile)

}
