% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryHasBarcodeInfo}
\alias{SequencingSummaryHasBarcodeInfo}
\title{does the SequencingSummary information define a barcode?}
\usage{
SequencingSummaryHasBarcodeInfo()
}
\value{
logical
}
\description{
does the SequencingSummary information define a barcode?
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
SequencingSummaryHasBarcodeInfo()

}
