% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{sequencingSummaryReadLengthQualityDensity}
\alias{sequencingSummaryReadLengthQualityDensity}
\title{plot a density map of sequence lengths and quality scores}
\usage{
sequencingSummaryReadLengthQualityDensity(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing densities of read length and quality distribution
}
\description{
plots the density plot of read length against read mean quality scores
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- sequencingSummaryReadLengthQualityDensity(seqsum)

}
