% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{sequencingSummaryReadQualityHistogram}
\alias{sequencingSummaryReadQualityHistogram}
\title{plot a histogram of sequence quality scores}
\usage{
sequencingSummaryReadQualityHistogram(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing read quality distribution
}
\description{
plots the histogram of read mean quality scores shaded by pass/fail status
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- sequencingSummaryReadQualityHistogram(seqsum)

}
