% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummarySpeedPlot}
\alias{SequencingSummarySpeedPlot}
\title{plot speed of sequencing against time (bases per second distribution)}
\usage{
SequencingSummarySpeedPlot(seqsum, scaling = 1, sampleHours = 48,
  sampleIntervalMinutes = 60)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{scaling}{scale factor for the data}

\item{sampleHours}{is the number of hours to plot data for (default is 48)}

\item{sampleIntervalMinutes}{is the resolution to plot data at}
}
\value{
ggplot2 showing temporal performance
}
\description{
plots a ggplot2 box-and-whisker plot for the distribution of sequencing speeds against time
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummarySpeedPlot(seqsum)
}

}
