% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{SequencingSummaryT50}
\alias{SequencingSummaryT50}
\title{calculates the timepoint within a sequencing run where 50percent of the data is produced}
\usage{
SequencingSummaryT50(seqsum, t = 0.5, scaling = 1, sampleHours = 48,
  sampleIntervalMinutes = 60)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{t}{is a fractional point through run against which time will be calculated}

\item{scaling}{factor}

\item{sampleHours}{is the number of hours to consider}

\item{sampleIntervalMinutes}{is the resolution of the plot in minutes}
}
\value{
a numeric value expressed in hours
}
\description{
an accessory method for identifying a timepoint where a given amount of data has been produced
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
T50 <- SequencingSummaryT50(seqsum)

}
