% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{sequencingSummaryWeightedReadLength}
\alias{sequencingSummaryWeightedReadLength}
\title{plot a weighted histogram of sequence read lengths}
\usage{
sequencingSummaryWeightedReadLength(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing weighted read length distribution
}
\description{
plots the histogram of read lengths, weighted, and shaded by pass/fail status
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- sequencingSummaryWeightedReadLength(seqsum)

}
