% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{setCachedYAMLValue}
\alias{setCachedYAMLValue}
\title{update the stored parameter value from named config field}
\usage{
setCachedYAMLValue(yaml = "config", field, value, subenv = NULL)
}
\arguments{
\item{yaml}{the name of the YAML object}

\item{field}{the parameter field to lookup}

\item{value}{to value to update}

\item{subenv}{the sub-environment to process}
}
\value{
TRUE or FALSE if the change can be made
}
\description{
update the stored parameter value from named config field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
referenceGenome <- system.file("extdata", "cas9_demo_ref.fasta", package = "nanopoRe")
setCachedYAMLValue(field="reference_genome", value=referenceGenome)
getCachedYAMLValue(field="reference_genome")

}
