% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{setCas9ParameterValue}
\alias{setCas9ParameterValue}
\title{update the stored parameter value from named config field}
\usage{
setCas9ParameterValue(field, value)
}
\arguments{
\item{field}{the parameter field to lookup}

\item{value}{to value to update}
}
\value{
TRUE or FALSE if the change can be made
}
\description{
update the stored parameter value from named config field
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
referenceGenome <- system.file("extdata", "cas9_demo_ref.fasta", package = "nanopoRe")
setCas9ParameterValue("reference_genome", referenceGenome)
getCas9ParameterValue("reference_genome")

}
