% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JupyterRpy2.R
\name{setLogFile}
\alias{setLogFile}
\title{method to mask R output by writing to a log file}
\usage{
setLogFile(filename = NULL, append = TRUE)
}
\arguments{
\item{filename}{the path to the logfile to use}

\item{append}{is a boolean defining whether the log file should be appended or ab initio}
}
\value{
None
}
\description{
creates a log file and writes R code and messages on STDERR to the log file
}
\examples{
\dontrun{
setLogFile("demofile.log", append=FALSE)
unsetLog()
}
}
