% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{setPlotDimensions}
\alias{setPlotDimensions}
\title{set dimension definitions for an allowed ggplot2 object}
\usage{
setPlotDimensions(dim)
}
\arguments{
\item{dim}{list of properties}
}
\value{
list defining dpi, width, height and units
}
\description{
This method sets and stores the list with the defined metrics for a ggplot2 presentation
}
\examples{
setPlotDimensions(list(dpi=180, width=21, height=12, units='cm'))
getPlotDimensions()

}
