% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{setReferenceGenome}
\alias{setReferenceGenome}
\title{set path for reference genome object}
\usage{
setReferenceGenome(reference.file)
}
\arguments{
\item{reference.file}{is the location of the fasta format sequence file for
reference genome}
}
\value{
None
}
\description{
This method defines the path where reference genome is stored
}
\examples{
init()
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
# and check that something has been stored ...
getReferenceGenome()

}
