% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Results.R
\name{setRpath}
\alias{setRpath}
\title{set path for R results}
\usage{
setRpath(path)
}
\arguments{
\item{path}{is the location to the directory for storing results}
}
\value{
None
}
\description{
This method sets a path where default R results will be stored during an analysis
}
\examples{
setRpath(file.path(tempdir(), "nanopoRe"))
getRpath()

}
