% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnifflesSV.R
\name{snifflesKaryogram}
\alias{snifflesKaryogram}
\title{prepare karyogram of annotated SVs}
\usage{
snifflesKaryogram(vcfFile, demoChr = NULL)
}
\arguments{
\item{vcfFile}{is the path to the vcfFile to plot}

\item{demoChr}{defines a dummy chromosome for plotting}
}
\value{
a ggbio/ggplot2 graph
}
\description{
This method will prepare a karyogram of annotated SVs against the genome reference
}
\examples{
vcfFile <- system.file("extdata", "GM24385.nf7.chr20.vcf", package = "nanopoRe")
# a reference genome should really be defined with setReferenceGenome
snifflesKaryogram(vcfFile, demoChr=20)

}
