% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{sourceCas9Parameters}
\alias{sourceCas9Parameters}
\title{import the cas9 parameters from YAML file}
\usage{
sourceCas9Parameters(yamlFile)
}
\arguments{
\item{yamlFile}{is a path to a stored YAML config file}
}
\value{
None
}
\description{
import the cas9 parameters from YAML file
}
\examples{
init()
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)

}
