% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnifflesSV.R
\name{svLengthDistribution}
\alias{svLengthDistribution}
\title{prepare figure of SV length distribution}
\usage{
svLengthDistribution(vcfFile)
}
\arguments{
\item{vcfFile}{is the path to the vcfFile to plot}
}
\value{
a ggplot2 graph
}
\description{
This method will prepare figure of SV length distribution
}
\examples{
vcfFile <- system.file("extdata", "GM24385.nf7.chr20.vcf",
    package = "nanopoRe")
svLengthDistribution(vcfFile)

}
