% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONT_tutorial_sv.R
\name{SVMappingCharacteristicTable}
\alias{SVMappingCharacteristicTable}
\title{ont_tutorial_sv = summarise the mapping characteristics from a given BAM file}
\usage{
SVMappingCharacteristicTable(bamSummary, validationResponse, bamFile)
}
\arguments{
\item{bamSummary}{is the result from  bamSummarise}

\item{validationResponse}{is the product of the floundeR fastqCheckup method}

\item{bamFile}{is a path to a file}
}
\value{
kable table of mapping characteristics as tuned for the ont_tutorial_sv document
}
\description{
this method prepares a kable format table summarising the key mapping characteritics from a provided
BAM file summary (from bamSummarise). This will describe the mapped sequence reads that are defined
as primary, secondary and supplementary and will overlay additional data that releases to read
length, read quality and other characteristics as reported in the tutorial document
}
\examples{
demoBam <- system.file("extdata",
    "Ecoli_zymo_R10_filt_subs.bam",
    package = "nanopoRe")
demoFastq <- system.file("extdata",
    "Ecoli_zymo_R10.fastq.gz",
    package = "nanopoRe")
referenceGenome <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceGenome)
loadReferenceGenome()
bamSummary <- bamSummarise(demoBam, force=FALSE, blockSize=1000L)
fastqView <- fastqCheckup(demoFastq)
kable <- SVMappingCharacteristicTable(bamSummary, fastqView, demoBam)

}
