% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{testBam}
\alias{testBam}
\title{extract content from a BAM file}
\usage{
testBam(bamFile)
}
\arguments{
\item{bamFile}{is the location to the BAM file to parse}
}
\value{
list of summary observations
}
\description{
This method will extract a block of BAM content from a given BAM file
}
\examples{
\dontrun{
testBam(file.path("Analysis", "Minimap2", "MyBamFile.bam"))
}

}
