% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JupyterRpy2.R
\name{unsetLog}
\alias{unsetLog}
\title{method to output capture to log file}
\usage{
unsetLog()
}
\value{
None
}
\description{
this assumes that a log file has been written to and ends the sinking
}
\examples{
\dontrun{
setLogFile("demofile.log", append=FALSE)
unsetLog()
}
}
