% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamParser.R
\name{bamSummarise}
\alias{bamSummarise}
\title{summarise mapping observations from a BAM file}
\usage{
bamSummarise(bamFile, force = FALSE, blockSize = 50000L)
}
\arguments{
\item{bamFile}{is the location to the BAM file to parse}

\item{force}{logical value describing whether the analysis should be force recalculated}

\item{blockSize}{the number of reads to process at the time as iterating through}
}
\value{
data.frame of per read summary observations
}
\description{
This method will parse a BAM file and summarise the mapping observations in a
way that can be used for preparation of figures and confidence plots by
the nanopoRe package
}
\examples{
demoBam <- system.file("extdata",
    "Ecoli_zymo_R10_filt_subs.bam",
    package = "nanopoRe")
bamSummary <- bamSummarise(demoBam, force=FALSE, blockSize=1000L)


}
