% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{cas9WriteExcelOnTarget}
\alias{cas9WriteExcelOnTarget}
\title{write summary of on-target mapping results to an excel format result file}
\usage{
cas9WriteExcelOnTarget()
}
\value{
None
}
\description{
write summary of on-target mapping results to an excel format result file
}
\examples{
\dontrun{
cas9WriteExcelOnTarget()
}
}
