% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{chromosomeMappingSummary}
\alias{chromosomeMappingSummary}
\title{prepare mapping summary information by chromosome}
\usage{
chromosomeMappingSummary(chrIds, bamFile, flag = "Primary")
}
\arguments{
\item{chrIds}{vector of chromosome ids}

\item{bamFile}{to the bamFile used for analysis}

\item{flag}{to define whether mapping is reported at the Primary level}
}
\value{
data.frame of mapping characteristics
}
\description{
This method will return a data.frame of basic per chromosome mapping statistics
}
\examples{
\dontrun{
chromosomeMappingSummary(c("1","2"), bamFile)
}

}
