% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastqValidator}
\alias{fastqValidator}
\title{parse a fastq file aiming to validate sequences}
\usage{
fastqValidator(fastq)
}
\arguments{
\item{fastq}{A fastq format DNA/RNA sequence file}
}
\value{
logical defining if fastq provided is indeed valid fastq
}
\description{
fastqValidator will parse the specified fastq (or fastq.gz) file looking for fastq
entry compliance. A boolean value of overall file compliance will be returned. Additional
summary counts describing the reason for rejection are also made available
}
\examples{
fastq <- system.file("extdata", "example.fastq.gz", package = "nanopoRe")
fastqValidator(fastq)
badFastq <- system.file("extdata", "frankenFastq.fastq.gz", package = "nanopoRe")
fastqValidator(badFastq)

}
