% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fixFastq}
\alias{fixFastq}
\title{fix a corrupted fastq file (if fastq-like)}
\usage{
fixFastq(fastq, newfastq)
}
\arguments{
\item{fastq}{file location of fastq source}

\item{newfastq}{location of file to write content to}
}
\description{
fix a corrupted fastq file (if fastq-like)
}
