% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceGenome.R
\name{getReferenceGenome}
\alias{getReferenceGenome}
\title{get path reference genome object}
\usage{
getReferenceGenome()
}
\value{
character representation of path to stored fasta format file object
}
\description{
This method returns the stored path for the reference genome object
}
\examples{
init()
referenceFasta <- system.file("extdata",
    "Escherichia_coli_complete_genome.fasta",
    package = "nanopoRe")
setReferenceGenome(referenceFasta)
# and check that something has been stored ...
getReferenceGenome()

}
