% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{ggbiosave}
\alias{ggbiosave}
\title{prepare a barchart showing number of bases per chromosome coloured by mapping-type}
\usage{
ggbiosave(filename, plot = last_plot(),
  device = default_device(filename), path = NULL, scale = 1,
  width = par("din")[1], height = par("din")[2], units = c("in",
  "cm", "mm"), dpi = 300, limitsize = TRUE, ...)
}
\arguments{
\item{filename}{filename}

\item{plot}{plot}

\item{device}{device}

\item{path}{path}

\item{scale}{scale}

\item{width}{width}

\item{height}{height}

\item{units}{units}

\item{dpi}{dpi}

\item{limitsize}{limitsize}

\item{...}{to downstream methods}
}
\value{
path to image
}
\description{
prepare a barchart showing number of bases per chromosome coloured by mapping-type;
enables looking for unexpected distributions of off-target, background etc per chromosome
}
