% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{importCas9TutorialData}
\alias{importCas9TutorialData}
\title{imports the R target output generated by the ont_tutorial_cas9 workflow}
\usage{
importCas9TutorialData(reference, study)
}
\arguments{
\item{reference}{is a pointer to the reference genome used}

\item{study}{is the named study used within the tutorial workflow}
}
\value{
None
}
\description{
a method for streamlining the import of R format data from the pre-calculated content
generated within the ont_tutorial_cas9 workflow. This is an accessory method to abstract some
of the code that looks out-of-control in e.g. the Epi2me-labs Jupyter notebooks
}
\examples{
\dontrun{
init()
}
}
