% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{importSequencingSummary}
\alias{importSequencingSummary}
\title{load a sequencing_summary.txt file into memory}
\usage{
importSequencingSummary(seqsum)
}
\arguments{
\item{seqsum}{is a path to a file}
}
\value{
data.frame of observations from the sequencing_summary.txt file provided
}
\description{
importSequencingSummary loads a sequencing_summary.txt file into
memory and performs basic sanity checking to ensure that the file
is cleaned of potential duplicate headers from aggregation
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)

}
