% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{is_casDataRun}
\alias{is_casDataRun}
\title{accessory method to test whether the cas9 data indexing has been performed}
\usage{
is_casDataRun()
}
\value{
TRUE or FALSE depending
}
\description{
The cas9 workflow involves a deep dive into the experimental data and produces
a couple of Rdata files that contain the experimental context and information
relating to on-target, off-target and other effects. This method tests whether
the analysis has been performed
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
is_casDataRun()

}
