% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentPlots.R
\name{LeftRightPlot}
\alias{LeftRightPlot}
\title{plot two ggplot2 figures side by side}
\usage{
LeftRightPlot(plotLeft, plotRight, ldim = 2, rdim = 2)
}
\arguments{
\item{plotLeft}{is the left-hand plot}

\item{plotRight}{is the right-hand plot}

\item{ldim}{relative scaling for left plot (2)}

\item{rdim}{relative scaling for right plot (2)}
}
\value{
a ggplot2 object
}
\description{
This method will plot 2 figures in a single plot space
}
\examples{
\dontrun{
LeftRightPlot(p1, p2)
}

}
