% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{listCachedObjects}
\alias{listCachedObjects}
\title{Initialise the NanopoRe environment}
\usage{
listCachedObjects()
}
\value{
None
}
\description{
Creates a NanopoRe environment; package specific parameters and values will be stored within this
environment; the name of the environment is defined internally
}
\examples{
\dontrun{
Nanopore::init()
}
}
