% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checksums.R
\name{md5sum}
\alias{md5sum}
\title{calculate md5sum checksum for qualified file}
\usage{
md5sum(filename)
}
\arguments{
\item{filename}{is a qualified path to a file of interest}
}
\value{
character representation of md5sum
}
\description{
This method calculates the md5 checksum for the specified file
}
\examples{
\dontrun{
md5sum(file.path('DESCRIPTION'))
}

}
