% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetEnrichment.R
\name{RunEnrichmentAnalysis}
\alias{RunEnrichmentAnalysis}
\title{perform an R-based enrichment analysis for a target region of the genome}
\usage{
RunEnrichmentAnalysis(force = FALSE, ...)
}
\arguments{
\item{force}{defines whether a reanalysis of the data should be forced
if results already existing}

\item{...}{- who knows - I'm using this for mc.cores}
}
\value{
TRUE or FALSE depending
}
\description{
This method is associated with the cas9-mediated enrichment of genomic
target regions ...
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
importConfigYAML(yamlFile=yamlFile)
isEnrichmentAnalysisComplete()
# this is a header to a document - not a live Snakemake workspace - demo
# how to actually run ad-hoc the code
setCachedYAMLValue(
    field="reference_genome",
    value=system.file("extdata", "cas9_demo_ref.fasta", package="nanopoRe"))
setCachedYAMLValue(
    field="target_regions",
    value=system.file("extdata", "cas9_demo_target.bed", package="nanopoRe"))
addCachedYAMLValue(
    field="bam_file",
    value=system.file("extdata", "cas9_FAK76554.bam", package ="nanopoRe"))
addCachedYAMLValue(
    field="unmapped_quals",
    value=system.file("extdata", "cas9_FAK76554.unmapped.quals",
    package="nanopoRe"))
# lines between the previous comment and here are typically not required
RunEnrichmentAnalysis(mc.cores=1)

}
