% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetEnrichment.R
\name{RunEnrichmentAnalysis}
\alias{RunEnrichmentAnalysis}
\title{perform an R-based enrichment analysis for a target region of the genome}
\usage{
RunEnrichmentAnalysis(force = FALSE)
}
\arguments{
\item{force}{defines whether a reanalysis of the data should be forced
if results already existing}
}
\value{
TRUE or FALSE depending
}
\description{
This method is associated with the cas9-mediated enrichment of genomic
target regions ...
}
\examples{
yamlFile <- system.file("extdata", "cas9_demo.yaml", package = "nanopoRe")
sourceCas9Parameters(yamlFile)
is_casDataRun()
RunEnrichmentAnalysis()

}
