% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequenceSummaryBarcodeHistogram}
\alias{SequenceSummaryBarcodeHistogram}
\title{present a histogram of read count by sorted barcode Id}
\usage{
SequenceSummaryBarcodeHistogram(seqsum = NA, bcthreshold = 150)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{bcthreshold}{the threshold number of reads for a barcode to be considered (150)}
}
\value{
a ggplot2 histogram or NULL
}
\description{
present a ggplot2 histogram of read counts by sorted barcode Id
}
\examples{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
importSequencingSummary(seqsumFile)
barcodeFile <- system.file('extdata', 'barcoding_summary.txt.bz2', package = 'nanopoRe')
SequencingSummaryBarcodeMerge(barcodeFile=barcodeFile)
SequenceSummaryBarcodeHistogram()

}
