% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequenceSummaryBarcodeInfoGraphic}
\alias{SequenceSummaryBarcodeInfoGraphic}
\title{presents an emojifont based infographic for barcode characteristics}
\usage{
SequenceSummaryBarcodeInfoGraphic(seqsum = NA, bcthreshold = 150)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{bcthreshold}{the threshold number of reads for a barcode to be considered (150)}
}
\value{
a numeric value expressed in gigabases
}
\description{
an accessory method for various logicals; simple fractional base calculator
}
\examples{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
importSequencingSummary(seqsumFile)
barcodeFile <- system.file('extdata', 'barcoding_summary.txt.bz2', package = 'nanopoRe')
SequencingSummaryBarcodeMerge(barcodeFile=barcodeFile)
SequenceSummaryBarcodeInfoGraphic()

}
