% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequencingSummaryBarcodeMerge}
\alias{SequencingSummaryBarcodeMerge}
\title{calculates the fractional number of bases according to supplied b parameter}
\usage{
SequencingSummaryBarcodeMerge(seqsum = NA, barcodeFile = NULL)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{barcodeFile}{pointer to a barcode file as produced by Guppy}
}
\value{
a data.frame with barcode information
}
\description{
an accessory method for various logicals; simple fractional base calculator
}
\examples{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
importSequencingSummary(seqsumFile)
barcodeFile <- system.file('extdata', 'barcoding_summary.txt.bz2', package = 'nanopoRe')
SequencingSummaryBarcodeMerge(barcodeFile=barcodeFile)

}
