% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummaryBarcodes.R
\name{SequencingSummaryBarcodeMerge}
\alias{SequencingSummaryBarcodeMerge}
\title{calculates the fractional number of bases according to supplied b parameter}
\usage{
SequencingSummaryBarcodeMerge(seqsum, barcodeFile)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{barcodeFile}{pointer to a barcode file as produced by Guppy}
}
\value{
a numeric value expressed in gigabases
}
\description{
an accessory method for various logicals; simple fractional base calculator
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
barcodeFile <- NULL
seqsum <- SequencingSummaryBarcodeMerge(seqsum, barcodeFile)
}

}
