% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryCumulativeReads}
\alias{SequencingSummaryCumulativeReads}
\title{plot cumulative volumes of sequence reads}
\usage{
SequencingSummaryCumulativeReads(seqsum, scaling = 1, sampleHours = 48,
  sampleIntervalMinutes = 60)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{scaling}{scale factor for the data}

\item{sampleHours}{is the number of hours to plot data for (default is 48)}

\item{sampleIntervalMinutes}{is the resolution to plot data at}
}
\value{
ggplot2 showing temporal performance
}
\description{
plots a ggplot2 graph of accumulated sequence reads against time for run and separates passed and
failed sequence reads
}
\examples{
\dontrun{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryCumulativeReads(seqsum)
}

}
