% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryPassGauge}
\alias{SequencingSummaryPassGauge}
\title{prepare a gauge plot of sequencing_summary reads passing QC}
\usage{
SequencingSummaryPassGauge(seqsum = NA)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 gauge plot
}
\description{
plots the basic eye-candy gauge plot of reads passing QC threshold
}
\examples{
\dontrun{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryPassGauge(seqsum)
}

}
