% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{sequencingSummaryReadLengthHistogram}
\alias{sequencingSummaryReadLengthHistogram}
\title{plot a histogram of sequence read lengths}
\usage{
sequencingSummaryReadLengthHistogram(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing read length distribution
}
\description{
plots the histogram of read lengths shaded by pass/fail status
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- sequencingSummaryReadLengthHistogram(seqsum)

}
