% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryReadQualityHistogram}
\alias{SequencingSummaryReadQualityHistogram}
\title{plot a histogram of sequence quality scores}
\usage{
SequencingSummaryReadQualityHistogram(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing read quality distribution
}
\description{
plots the histogram of read mean quality scores shaded by pass/fail status
}
\examples{
init()
seqsumFile <- system.file("extdata",
    "sequencing_summary.txt.bz2", package = "nanopoRe")
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryReadQualityHistogram(seqsum)

}
