% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSummary.R
\name{SequencingSummaryTemporalThroughput}
\alias{SequencingSummaryTemporalThroughput}
\title{plot a sequence throughput against time for specified sequencing_summary run}
\usage{
SequencingSummaryTemporalThroughput(seqsum, scaling = 1,
  sampleHours = 48, sampleIntervalMinutes = 60)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}

\item{scaling}{scale factor for the data}

\item{sampleHours}{is the number of hours to plot data for (default is 48)}

\item{sampleIntervalMinutes}{is the resolution to plot data at}
}
\value{
ggplot2 showing temporal performance
}
\description{
plots a ggplot2 graph of performance against time for run and separates passed and failed sequence reads
}
\examples{
seqsumFile <- system.file("extdata", "sequencing_summary.txt.bz2", package = "nanopoRe", mustWork = TRUE)
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryTemporalThroughput(seqsum)

}
