% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSummary.R
\name{SequencingSummaryWeightedReadLength}
\alias{SequencingSummaryWeightedReadLength}
\title{plot a weighted histogram of sequence read lengths}
\usage{
SequencingSummaryWeightedReadLength(seqsum)
}
\arguments{
\item{seqsum}{is the data.frame object as prepared by importSequencingSummary}
}
\value{
ggplot2 showing weighted read length distribution
}
\description{
plots the histogram of read lengths, weighted, and shaded by pass/fail status
}
\examples{
\dontrun{
seqsumFile <- system.file('extdata', 'sequencing_summary.txt.bz2', package = 'nanopoRe')
seqsum <- importSequencingSummary(seqsumFile)
plot <- SequencingSummaryWeightedReadLength(seqsum)
}

}
