% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas9_tutorial.R
\name{sourceCas9Parameters}
\alias{sourceCas9Parameters}
\title{import the cas9 parameters from YAML file}
\usage{
sourceCas9Parameters(yaml)
}
\arguments{
\item{yaml}{is a yaml object as prepared by loading from file}
}
\value{
None
}
\description{
import the cas9 parameters from YAML file
}
\examples{
\dontrun{
myYaml <- yaml.load_file("config.yaml")
sourceCas9Parameters(myYaml)
}
}
