% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnifflesSV.R
\name{Vcf2GRanges}
\alias{Vcf2GRanges}
\title{convert VCF content into a GRanges object for nanopoRe usage}
\usage{
Vcf2GRanges(vcfFile)
}
\arguments{
\item{vcfFile}{is the path to the vcfFile to plot}
}
\value{
a ggbio/ggplot2 graph
}
\description{
This method will prepare a karyogram of annotated SVs against the genome reference
}
\examples{
\dontrun{
Vcf2GRanges(vcfFile)
}

}
